float weight

+++ src/models/NewWorkflowActionConditionalModel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Integer, String, JSON
+from sqlalchemy import Boolean, Float, Integer, String, JSON
 from sqlalchemy.orm import Mapped, mapped_column
 from ._base import Base
 
@@ -10,7 +10,7 @@ class BaseNewWorkflowActionConditionalModel(Base):
     workflow_action_id: Mapped[str] = mapped_column(String, nullable=True, index=True)
     type: Mapped[str] = mapped_column(String, nullable=True, index=True)
     order: Mapped[int] = mapped_column(Integer, nullable=True, index=True)
-    weight: Mapped[int] = mapped_column(Integer, nullable=True, index=False)
+    weight: Mapped[float] = mapped_column(Float, nullable=True, index=False)
     variable: Mapped[str] = mapped_column(String, nullable=True, index=False)
     value: Mapped[str] = mapped_column(String, nullable=True, index=False)
     metadata_value: Mapped[str] = mapped_column(String, nullable=True, index=False)
